+++ Tests/Fixtures/RetentionFixtures/testRetainsDynamicMemberLookupSubscript.swift
@@ -1,0 +1,10 @@
+@dynamicMemberLookup
+public struct FixtureStruct7 {
+    subscript(dynamicMember member: String) -> String {
+        ""
+    }
+
+    subscript(other: String) -> String {
+        ""
+    }
+}